+++ tests/testthat/test-pkg-fixest.R
@@ -3,6 +3,17 @@ skip_if_not_installed("fixest", minimum_version = "0.10.5")
 requiet("fixest")
 
 
+
+test_that("issue 591", {
+  requiet("data.table")
+  dt <- as.data.table(mtcars)
+  mod1 <- feols(c(mpg, disp) ~ hp + drat + qsec, data = mtcars, subset = mtcars$gear > 3)
+  mod2 <- feols(c(mpg, disp) ~ hp + drat + qsec, data = dt, subset = dt[, gear > 3])
+  expect_equal(ncol(modelsummary(mod1, "dataframe")), 5)
+  expect_equal(ncol(modelsummary(mod2, "dataframe")), 5)
+})
+
+
 test_that("fixest built in SE label", {
     mod <- feols(Euros ~ dist_km | Product + Destination + Origin, data = trade)
     tab <- modelsummary(mod, "data.frame")