+++ NEWS.md
@@ -7,7 +7,7 @@ Breaking change:
 New features:
 
 * `fmt` can round to significant digits on a per-term basis with `fmt="s1"`, `fmt="s2"`, etc.
-* `fmt` can format different parameters differently with a named list:
+* `fmt` can format parameters differently with a named list:
   - `fmt=list("(Intercept)" = 2, "hp" = 3)`
   - `fmt=list("estimate" = 2, "std.error" = 3)`
 * New styles for default column labels in `modelsummary`, such as Roman Numerals or letters in parentheses.